router: move context params stuff to router.go
