--- params.go
@@ -1,32 +1,0 @@
-package httprouter
-
-import (
-	"context"
-	"net/http"
-)
-
-type paramsKey struct{}
-
-// ParamsKey is the request context key under which URL params are stored.
-var ParamsKey = paramsKey{}
-
-// Handler is an adapter which allows the usage of an http.Handler as a
-// request handle.
-// The Params are available in the request context under ParamsKey.
-func (r *Router) Handler(method, path string, handler http.Handler) {
-	r.Handle(method, path,
-		func(w http.ResponseWriter, req *http.Request, p Params) {
-			ctx := req.Context()
-			ctx = context.WithValue(ctx, ParamsKey, p)
-			req = req.WithContext(ctx)
-			handler.ServeHTTP(w, req)
-		},
-	)
-}
-
-// ParamsFromContext pulls the URL parameters from a request context,
-// or returns nil if none are present.
-func ParamsFromContext(ctx context.Context) Params {
-	p, _ := ctx.Value(ParamsKey).(Params)
-	return p
-}
